feat: symphony extension settings now extracted into its own component
